refactor: isolated Docker management and scene finder utilities for improved functionality and better readability
